+++ lib/flows/locales/en/marriage-abroad.yml
@@ -31,20 +31,25 @@ en-GB:
           You may be able to get married at the British embassy or consulate in %{country_name_lowercase_prefix}.
         documents_needed_ss_marriage: |
           ##What documents you’ll need
+
           You’ll need proof that you’ve been resident in the area covered by the embassy or consulate for at least 7 days, like an employer’s letter or a bank statement.
+
           You’ll both need your original passports. If either of you have been divorced, widowed or in a civil partnership before, you’ll also need either:
 
           - a [decree absolute or final order](/copy-decree-absolute-final-order)
           - the [death certificate](/order-copy-birth-death-marriage-certificate/)
         what_to_do_ss_marriage: |
           ##What you need to do
+
           Once you’ve made your appointment, the embassy or consulate will give you:
 
           - a notice of registration
           - a declaration that you and your partner will need to swear, stating that you’re legally entitled to marry
 
           Once you’ve submitted these and paid the registration fee (see the fees table on this page), the embassy or consulate will display your notice publicly for 14 days.
+
           As long as nobody registers an objection during the 14 days, the registration officer can then register your marriage any time until 3 months after you gave notice.
+
           You’ll need to provide two witnesses and pay a fee to register your marriage. You’ll need to pay an additional fee for your marriage certificate (see the fees table on this page).
         fees_table_ss_marriage: |
           ##Fees
@@ -54,29 +59,42 @@ en-GB:
           Posting a marriage notice | £65
           Performing a marriage ceremony | £140
           Issuing a marriage certificate | £65
+        fees_table_ss_marriage_alt: |
+          ##Fees
+
+          Service | Fee
+          -|-
+          Receiving a notice of registration  | £65
+          Registering a marriage | £140
+          Issuing a marriage certificate | £65
 #same sex and civil partnership phrases
         title_ss_marriage_and_partnership: |
           Same-sex marriage and civil partnership
         able_to_ss_marriage_and_partnership: |
-          You may be able to register a civil partnership or get married
-        at_embassy_or_consulate: at the British embassy or consulate in %{country_name_lowercase_prefix}.
+          You may be able to register a civil partnership or get married at the British embassy or consulate in %{country_name_lowercase_prefix}.
 
         documents_needed_ss_marriage_and_partnership: |
           ##What documents you’ll need
+
           You’ll need proof that you’ve been resident in the area covered by the embassy or consulate for at least 7 days, like an employer’s letter or a bank statement.
+
           You’ll both need your original passports. If either of you have been divorced, widowed or in a civil partnership before, you’ll also need either:
 
           -  a [decree absolute or final order](/copy-decree-absolute-final-order)
           -  the [death certificate](/order-copy-birth-death-marriage-certificate/)
 
         what_to_do_ss_marriage_and_partnership: |
+          ##What you need to do
+
           Once you’ve made your appointment, the embassy or consulate will give you:
 
           -  a notice of registration
           -  a declaration that you and your partner will need to swear, stating that you’re legally entitled to marry or enter into a civil partnership
 
           Once you’ve submitted these and paid the registration fee (see the fees table on this page), the embassy or consulate will display your notice publicly for 14 days.
+
           As long as nobody registers an objection during the 14 days, the registration officer can then register your marriage or partnership any time until 3 months after you gave notice.
+
           You’ll need to provide 2 witnesses and pay a fee to register your marriage or civil partnership. You’ll need to pay an additional fee for your marriage or civil partnership certificate (see the fees table on this page).
         fees_table_ss_marriage_and_partnership: |
           ##Fees
@@ -460,15 +478,15 @@ en-GB:
         cni_posted_if_no_objection_14_days: |
           You can normally get a CNI by giving a notice of marriage at your local register office or registrar in the UK. Find your local office or registrar in [England and Wales](/register-offices), [Scotland](http://www.gro-scotland.gov.uk/files1/registration/reglist.pdf) or [Northern Ireland](http://www.nidirect.gov.uk/index/information-and-services/government-citizens-and-rights/births-and-registration/district-registrars-in-northern-ireland.htm).
 
-            ^A CNI issued in England, Wales or Northern Ireland is valid for 6 months. A CNI issued in Scotland is valid for 3 months.^
+          ^A CNI issued in England, Wales or Northern Ireland is valid for 6 months. A CNI issued in Scotland is valid for 3 months.^
 
-            They’ll post your notice, and as long as nobody has registered an objection after 14 days, they’ll issue your CNI.
+          They’ll post your notice, and as long as nobody has registered an objection after 14 days, they’ll issue your CNI.
         cni_posted_if_no_objection_7_days: |
           You can normally get a CNI by giving a notice of marriage at your local register office or registrar in the UK. Find your local office or registrar in [England and Wales](/register-offices), [Scotland](http://www.gro-scotland.gov.uk/files1/registration/reglist.pdf) or [Northern Ireland](http://www.nidirect.gov.uk/index/information-and-services/government-citizens-and-rights/births-and-registration/district-registrars-in-northern-ireland.htm).
 
-            ^A CNI issued in England, Wales or Northern Ireland is valid for 6 months. A CNI issued in Scotland is valid for 3 months.^
+          ^A CNI issued in England, Wales or Northern Ireland is valid for 6 months. A CNI issued in Scotland is valid for 3 months.^
 
-            They’ll post your notice, and as long as nobody has registered an objection after 7 days, they’ll issue your CNI.
+          They’ll post your notice, and as long as nobody has registered an objection after 7 days, they’ll issue your CNI.
         scotland_ni_resident_partner_irish_os_consular_cni_three: |
           You can normally get a CNI by giving a notice of marriage at local registrar if you live in [Scotland](http://www.gro-scotland.gov.uk/files1/registration/reglist.pdf) or [Northern Ireland](http://www.nidirect.gov.uk/index/information-and-services/government-citizens-and-rights/births-and-registration/district-registrars-in-northern-ireland.htm).
 
@@ -1670,4 +1688,8 @@ en-GB:
         title: "%{ss_title} in %{country_name_lowercase_prefix}"
         body: |
           %{ss_ceremony_body}
+      outcome_ss_marriage_not_possible:
+        title: Same-sex marriage and civil partnership in %{country_name_lowercase_prefix}
+        body: |
+          It’s not possible to get legal recognition for your same-sex relationship in %{country_name_lowercase_prefix}.
